==> notebooks/1 - Data Aquisition.ipynb <==
%%capture

%cd ..
%load_ext autoreload
%autoreload 2

from pathlib import Path
import pandas as pd
import numpy as np
import json

# Data Acquisition and Preparation

RAW_DATA_FOLDER = Path('data/raw/')
INTERMEDIATE_DATA_FOLDER = Path('data/interim/')
REFERENCE_FOLDER = Path('references/')

## Downloading Data

# TODO: Data Acquisition
# Salvar no RAW_DATA_FOLDER
