==> notebooks/1 - Data Preparation.ipynb <==
import json
from pathlib import Path
import pandas as pd
import numpy as np

pd.set_option('expand_frame_repr', True)
pd.set_option('display.column_space', 6)
pd.set_option('display.max_colwidth', 20)
pd.set_option('display.max_columns', None)
pd.set_option('display.show_dimensions', True)

# Data Preparation

The intent of this notebook is to tabulated the raw data from multiple files into one CSV file. The new data will be avaliable at the path defined in the variable `INTERMEDIATE_DATA_FOLDER` below.

CURRENT_WORK_DIRECTORY = Path().resolve().parents[0]

RAW_DATA_FOLDER = CURRENT_WORK_DIRECTORY / 'data/raw/'
INTERMEDIATE_DATA_FOLDER = CURRENT_WORK_DIRECTORY / 'data/interim/'
FULL_DATASET_PATH = INTERMEDIATE_DATA_FOLDER/ "fake_true_news.csv"

FAKE_DATA_FOLDER = RAW_DATA_FOLDER / 'fake'
TRUE_DATA_FOLDER = RAW_DATA_FOLDER / 'true'
FAKE_META_FOLDER = RAW_DATA_FOLDER / 'fake-meta-information'
TRUE_META_FOLDER = RAW_DATA_FOLDER / 'true-meta-information'

INTERMEDIATE_DATA_FOLDER.mkdir(exist_ok=True, parents=True)

## Preparing text data

First, we are going to load the full texts into a pandas dataframe for both fake and true news (indexed by the number of the file).

def create_text_df(folder):
    df_dict = {}
    for filepath in folder.glob("*.txt"):
        with open(filepath, 'r', encoding='utf-8') as f:
            df_dict[filepath.stem] = f.read() 
    return pd.DataFrame.from_dict(df_dict, orient='index', columns=['text'])

fake_text_df = create_text_df(FAKE_DATA_FOLDER)
true_text_df = create_text_df(TRUE_DATA_FOLDER)

display(fake_text_df.shape)
fake_text_df.head()

display(true_text_df.shape)
true_text_df.head()

## Metadata Dataset

On this section the metadata available as jsons are all tabulated into 2 dataframes (for both classes). The description for each variable (extracted from the source) can be found in *references/news_data_dictionary.json* file on the root of this repository.

def create_metadata_df(folder, metadata_columns):
    df_dict = {}
    df_dict = {k:[] for k in metadata_columns}
    df_dict["index"] = []
    
    for filepath in list(folder.glob("*.txt")):
        with open(filepath, 'r') as f:    
            df_dict["index"].append(filepath.stem.split("-")[0])
            for col, value in zip(metadata_columns, f.readlines()):
                df_dict[col].append(value[0:-1])
    
    df = pd.DataFrame(df_dict)
    df.index.name = None
    return df

metadata_columns = [
    "author", "link", "category", "date_of_publication",
    "tokens", "words_no_punctuation", "types", "links_inside", 
    "upper_words", "verbs", "subjuntive_imperative_verbs",
    "nouns", "adjectives", "adverbs", "modal_verbs", 
    "singular_first_second_personal_pronouns",
    "plural_first_personal_pronouns", "pronouns",
    "pausality", "characters", "average_sentence_length",
    "average_word_lenght", "percentage_spelling_errors",
    "emotiveness", "diversity"
]

fake_metadata_df = create_metadata_df(FAKE_META_FOLDER, metadata_columns)
true_metadata_df = create_metadata_df(TRUE_META_FOLDER, metadata_columns)

display(fake_metadata_df.shape)
fake_metadata_df.head()

display(true_metadata_df.shape)
true_metadata_df.head()

Some of the null values are defined as a "None" string and can difficult the next steps. We need to fix this behavior by replacing the string with np.nan

print("# of 'None' strings")
print(f'True: {len(fake_metadata_df[fake_metadata_df.isin(["None"]).any(axis=1)])}')
print(f'True: {len(true_metadata_df[true_metadata_df.isin(["None"]).any(axis=1)])}')

fake_metadata_df = fake_metadata_df.replace("None", np.nan)
true_metadata_df = true_metadata_df.replace("None", np.nan)

We also need to correct the data types for the variables. These types were defined manually. **NOTE:** the variable *date_of_publication* have a lot of different formats that could not be identified by the pandas library, since we are not going to use this data it will be loaded as a string.

metadata_dtypes = {
    "author": "string", "link": "string", "category": "string",
    "date_of_publication": "string",
    "tokens": "float", "words_no_punctuation": "float",
    "types": "float","links_inside": "float", "upper_words": "float",
    "verbs": "float", "subjuntive_imperative_verbs": "float", "nouns": "float", 
    "adjectives": "float", "adverbs": "float","modal_verbs": "float", 
    "singular_first_second_personal_pronouns": "float",
    "plural_first_personal_pronouns": "float", "pronouns": "float","characters": "float",
    "pausality": "float", "average_sentence_length": "float",
    "average_word_lenght": "float", "percentage_spelling_errors": "float",
    "emotiveness": "float", "diversity": "float"
}

fake_metadata_df = fake_metadata_df.astype(metadata_dtypes, errors='raise').set_index("index", drop=True)
true_metadata_df = true_metadata_df.astype(metadata_dtypes, errors='raise').set_index("index", drop=True)

## Creating full dataset

The only thing left is to merge the data into one unique csv. This data will be available at the `FULL_DATASET_PATH` at the beginning of this notebook.

### Fake News Dataset

fake_df = pd.concat([fake_text_df, fake_metadata_df], axis=1, sort=False)
fake_df.index = fake_df.index.astype(int)
fake_df = fake_df.reset_index().rename(columns={"index": "file_index"})
fake_df = fake_df.sort_index()

fake_df.shape

### True News Dataset

true_df = pd.concat([true_text_df, true_metadata_df], axis=1, sort=False)
true_df.index = true_df.index.astype(int)
true_df = true_df.reset_index().rename(columns={"index": "file_index"})
true_df = true_df.sort_index()

true_df.shape

### Full Dataset

result = pd.concat([true_df, fake_df], keys=['True', 'Fake'])
result = result.reset_index(level=0)
result = result.rename(columns={"level_0": "class"})
result.index.name = None

display(result.shape)
display(result.head())

result.to_csv(FULL_DATASET_PATH, index=False)
