==> notebooks/3 - Feature Engineering.ipynb <==
from pathlib import Path
import pandas as pd
import numpy as np
import itertools

# Feature Engineering

This notebook shows the steps taken for the feature engineering after the cleaning done on "2 - Cleaning"

CURRENT_WORK_DIRECTORY = Path().resolve().parents[0]
PROCESSED_DATA_FOLDER = CURRENT_WORK_DIRECTORY / 'data/processed/'

## Loading Data

data_df = pd.read_csv(PROCESSED_DATA_FOLDER/"fake_true_news_cleaned.csv", index_col=0)
data_df["class"] = data_df["class"].astype('category')

display(data_df.shape)
data_df.head()

## Ratio Variables

To minimize the bias related to true news being bigger than fake ones our approach is to calculate the ratio of some variables in respect of the number of tokens.

def calculate_ratio(s1, s2):
    return (s1 / s2) * 100      

data_df['percent_words_no_punctuation'] = calculate_ratio(data_df.words_no_punctuation, data_df.tokens)
data_df['percent_adjectives'] = calculate_ratio(data_df.adjectives, data_df.tokens)
data_df['percent_adverbs'] = calculate_ratio(data_df.adverbs, data_df.tokens)
data_df['percent_nouns'] = calculate_ratio(data_df.nouns, data_df.tokens)
data_df['percent_subjuntive_imperative_verbs'] = calculate_ratio(data_df.subjuntive_imperative_verbs, data_df.tokens)
data_df['percent_upper_words'] = calculate_ratio(data_df.upper_words, data_df.tokens)
data_df['percent_verbs'] = calculate_ratio(data_df.verbs, data_df.tokens)
data_df['percent_modal_verbs'] = calculate_ratio(data_df.modal_verbs, data_df.tokens)
data_df['percent_singular_first_second_personal_pronouns'] = calculate_ratio(data_df.singular_first_second_personal_pronouns, data_df.tokens)
data_df['percent_plural_first_personal_pronouns'] = calculate_ratio(data_df.plural_first_personal_pronouns, data_df.tokens)
data_df['percent_pronouns'] = calculate_ratio(data_df.pronouns, data_df.tokens)

data_df = data_df.drop(columns=[
    'words_no_punctuation', 'adjectives', 'adverbs', 'nouns', 
    'subjuntive_imperative_verbs', 'upper_words', 'verbs', 'modal_verbs', 'pronouns',
    'singular_first_second_personal_pronouns', 'plural_first_personal_pronouns' 
])

## Exporting Data

data_df.to_csv(PROCESSED_DATA_FOLDER/"fake_true_news_cleaned_features.csv")
