==> notebooks/5 - Modeling.ipynb <==
# Logistic Regression

#### The idea is to adjust the data on a logistic regression model

#Limpando o console
rm(list=ls())

#Packages
library(mfx)
library(stargazer) 
library(car)

#### Data

FILE_PATH = ".../data/fake_true_news_cleaned_features.csv"
Dados <- read.csv(FILE_PATH, sep=",")
#links_inside e upper_words

# data Classes
t(t(sapply(Dados, class)))

# Transformation in the variable class(Fake-true) in categoric
Dados$class <- as.factor(Dados$class)
class(Dados$class)

# Modeling

##### We utilized general linear models (glm) with Class as the dependent variable. The model's family are (family = binomial(link="logit"))

 We adjusted the model with all the variables of the data. According to the results 5 variables show statistical significance: : Types, characters, average_sentence_length, percent_subjuntive_imperative_verbs e percent_modal_verbs. Because of multicolinearity we adjusted the models without the Types and Character variables.

#### <font color=red> MODEL 1</font>

#Ajustando a regressão
my_model <- glm(class ~ file_index +  tokens + types + pausality + characters + average_sentence_length + average_word_lenght
                                 + percentage_spelling_errors + emotiveness + diversity + percent_words_no_punctuation +  percent_adjectives 
                                 + percent_adverbs + percent_nouns + percent_subjuntive_imperative_verbs + percent_upper_words + percent_verbs
                                 + percent_modal_verbs + percent_pronouns, data = Dados, family = binomial(link="logit"))
summary(my_model)

#modelo com variaveis significativas
my_model <- glm(class ~  types + characters + average_sentence_length
                                  +  percent_subjuntive_imperative_verbs
                                 + percent_modal_verbs, data = Dados, family = binomial(link="logit"))
summary(my_model)

#Previsao dos modelos
predictions_mlg <- predict(my_model, Dados, type="response")
predicted.classes <- ifelse(predictions_mlg  > 0.5, "1", "0")
       
#acuracia
acuracia_rate_mlg <- mean(predicted.classes == Dados$class)
print(acuracia_rate_mlg)

#### <font color=red> MODEL 2 </font>

#Ajustando a regressão
my_model <- glm(class ~    types  + characters +average_sentence_length 
                                + percent_subjuntive_imperative_verbs + percent_modal_verbs 
                              , data = Dados, family = binomial(link="logit"))
summary(my_model)

#Previsao dos modelos
predictions_mlg <- predict(my_model, Dados, type = "response")
predicted.classes <- ifelse(predictions_mlg  > 0.6, "1", "0")

#acuracia
acuracia_rate_mlg <- mean(predicted.classes == Dados$class)
print(acuracia_rate_mlg)

df <- table(Dados$class, predictions_mlg > 0.5) 
df

#Calculando a predição, Recall e F_measure

#Precision
precision <- df[1,1]/sum(df[1,1:2])
round(precision, 4)


#Recall
recall <- df[1,1]/sum(df[1:2,1])
round(recall,4)

#F-measure
Fmeasure <- 2 * precision * recall / (precision + recall)
round(Fmeasure,4)



#### Calculating the exponentiated coefficient for each model variable and their confidence intervals

exp(cbind(OR = coef(my_model), confint(my_model)))

#### <font color=red> MODEL 3 </font>

#Ajustando a regressão
my_model <- glm(class ~ characters +average_sentence_length 
                                + percent_subjuntive_imperative_verbs + percent_modal_verbs 
                              , data = Dados, family = binomial(link="logit"))
summary(my_model)

exp(cbind(OR = coef(my_model), confint(my_model)))

#Previsao dos modelos
predictions_mlg <- predict(my_model, Dados, type="response")
predicted.classes <- ifelse(predictions_mlg  > 0.6, "1", "0")
        
#acuracia
acuracia_rate_mlg <- mean(predicted.classes == Dados$class)
print(acuracia_rate_mlg)

#### <font color=red> MODEL 4 </font>

#Ajustando a regressão
my_model<-glm(class~types+average_sentence_length+percent_subjuntive_imperative_verbs+percent_modal_verbs,
              data = Dados, family = binomial(link="logit"))

summary(my_model)

#Previsao dos modelos
    predictions_mlg <- predict(my_model, Dados, type="response")
    predicted.classes <- ifelse(predictions_mlg  > 0.6, "1", "0")
        
#acuracia
    acuracia_rate_mlg <- mean(predicted.classes == Dados$class)
    print(acuracia_rate_mlg)

#Data
dat <- read.csv("C:\\Users\\Wanessa\\Desktop\\Regression Logist\\FAKE_R.csv", sep=",")
str(dat)

# Transformation in the variable class(Fake-true) in categoric
dat$class <- as.factor(dat$class)
class(dat$class)

#Dimension 
dim(dat)

#Class Variable
t(t(sapply(dat, class)))



# Train and Test

# 30 times 10 fold

#Captura o tamanho da amostra
n <- nrow(dat)

#30 replicas de 10-folds (Colocar qd terminar os valores certos)
R <- 30
folds <- 10

#mean_error_rate <- rep(0, R)
mean_erro_rate_mlg <- rep(0, R)
mean_acuracia_rate_mlg <- rep(0, R)
#error_rate <- rep(0, folds)
acuracia_rate_mlg <- rep(0, folds)
erro_rate_mlg <- rep(0, folds)

#Gerando os Dados
#fixando uma semente
se1 = 1994 ; se2 = 1991#random seed
set.seed(c(se1,se2), kind="Marsaglia-Multicarry")


for(r in 1 : R){
  for(i in 1 : folds){
    s <- sample(1 : n, replace = F, floor(.1*n))
    length(s)

    train <- dat[-s, ]
    test <- dat[s, ]
    #tirei isso para ficar igual
    #test <- dat[s, -ncol(dat)]
    target_test <- dat[s, ncol(dat)]
    dim(train)
    length( target_test)
    dim(test)

    #Aplico o modelo 1
    fit_mlg <- glm(class ~ types + characters + average_sentence_length + percent_subjuntive_imperative_verbs
                  + percent_modal_verbs, data = train, family = binomial(link="logit"))

    # coefficients
    #coef(fit_mlg)
    #Aplico o modelo 1
    #fit_mlg <- glm(class ~ types + characters + percent_modal_verbs, data = train, family = binomial(link="logit"))

    summary(fit_mlg)


    #Previsao dos modelos
    predictions_mlg <- predict(fit_mlg, test, type="response")
    head(predictions_mlg )

    #probabilities <- fit_mlg %>% predict(test, type = "response")
    #head(probabilities)

    contrasts(train$class)

    predicted.classes <- ifelse(predictions_mlg  > 0.5, "1", "0")
    head(predicted.classes)

    #acuracia
    acuracia_rate_mlg[i] <- mean(predicted.classes == test$class)
    acuracia_rate_mlg

    #taxa de errro
    #erro_rate_mlg[i] <- 1 - acuracia_rate_mlg
    #erro_rate_mlg
    #length(predicted.classes)


    #Matrizes de confusao
    #confusion_matrices_mlg <- table(predictions_mlg, target_test)


    #Calcula a taxa de erro dos classificadores
    #acuracia_rate_mlg[i] <- sum(diag(confusion_matrices_mlg))/nrow(test)
    #acuracia_rate_mlg <- sum(diag(confusion_matrices_mlg))/nrow(test)
    #acuracia_rate_mlg
    #error_rate_mlg <- 1 -  sum(diag(confusion_matrices_mlg))/nrow(test)
    #print(error_rate_mlg)
  }
  #Armazena a taxa de erro media dos folds para cada replica
  mean_acuracia_rate_mlg[r] <- mean(acuracia_rate_mlg)
  #mean_erro_rate_mlg[r] <- mean(erro_rate_mlg)
}


#mean(mean_error_rate)
mean(mean_acuracia_rate_mlg)
#mean(mean_erro_rate_mlg)
